runtime: address PR review comments (B2 + H1 H3 + M1 M3 M4 + misc)
Round of fixes off the latest PR review feedback:

Blocking
- Initialize ``_implicit_cache_handle`` in ``TorchTensorRTModule.__init__``
  so the slot exists on every construction path, not just ``setup_engine``.
  Drops a regression where ``set_extra_state`` (post-load) skipped the
  init and any subsequent ``mod.runtime_settings = ...`` raised
  AttributeError. Removes the matching ``# type: ignore[has-type]``.

High-priority
- ``_to_torchbind_handle`` now rejects a mixed-runtime case loudly:
  a Python ``RuntimeCacheHandle`` with a live pybind ``IRuntimeCache``
  but no torchbind sibling crossing into the C++ runtime path would
  silently orphan the cache.
- Also gates the str -> torchbind path on a truthy-string check
  (mirrors ``_materialize_implicit_handle``) so ``""`` doesn't
  construct a no-op torchbind handle.
- ``runtime_settings.setter`` now reconciles the handle that
  ``_dispatch_runtime_settings_to_engine`` substituted in, matching the
  ``setup_engine`` post-condition (``self._runtime_settings`` agrees
  with what the engine actually saw).

Medium
- ``runtime`` Bazel cc_library now exports ``TRTRuntimeConfig.h``
  alongside ``TRTEngine.h`` for symmetry with ``runtime_base`` and the
  ``include_files`` filegroup.
- ``_RuntimeCacheContextManager.__exit__`` synchronizes CUDA before
  save (avoids the detach-then-save race against a concurrent execute),
  and wraps save in try/except+warn so a transient filesystem failure
  on exit can't mask the with-block's actual exception.
- ``to_*_strategy`` now take ``int64_t`` and bounds-check before
  narrowing, so an out-of-range Python caller can't slip past the check
  via a silent ``int32_t`` overflow.

Low
- Drop the explicit local_defines select on ``ENABLE_FEATURE_DISABLE_RUNTIME_ALLOCATION``
  in ``core/runtime/BUILD`` (no longer needed on TRT-RTX 1.5+).
- ``.size()`` -> ``std::size(...)`` on the std::array uses in
  ``RuntimeSettings.cpp`` for consistency.
- Trim docstrings on ``enable_cudagraphs`` and ``set_cuda_graph_strategy``
  -- composition guidance moves to the runtime-settings doc page.
- Drop ``_NON_RTX_WARNING_EMITTED`` module global; emit unconditionally
  so test isolation can assert on it. Users wanting once-per-process
  behavior filter via ``warnings.simplefilter("once", UserWarning)``.
- Document multi-process lost-update semantics of the default
  ``runtime_cache`` path in the dataclass docstring.
- Skip the ``num_execution_contexts_created == 0`` assertion for NCCL
  engines in ``test_one_context_create_with_default_settings``: NCCL
  engines eagerly bind the comm at setup, which materializes the
  context. Comment in ``_TRTEngine`` flags the same divergence.

+++ core/runtime/RuntimeSettings.h
@@ -72,8 +72,8 @@ enum class CudaGraphStrategy : int32_t {
 // the enum (or throw with a clear message on out-of-range). Used only inside
 // the torchbind ``update_runtime_settings`` lambda -- the rest of the code
 // passes the enum type directly.
-[[nodiscard]] DynamicShapesKernelSpecializationStrategy to_dynamic_shapes_kernel_strategy(int32_t v);
-[[nodiscard]] CudaGraphStrategy to_cuda_graph_strategy(int32_t v);
+[[nodiscard]] DynamicShapesKernelSpecializationStrategy to_dynamic_shapes_kernel_strategy(int64_t v);
+[[nodiscard]] CudaGraphStrategy to_cuda_graph_strategy(int64_t v);
 
 // Per-engine runtime-only knobs sampled at IExecutionContext creation.
 //
